Securing the view so user must have an account to edit or delete a story

+++ stories/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Stories
 from .forms import StoryForm
 
@@ -29,9 +30,13 @@ def story_detail(request, story_id):
 
     return render(request, 'stories/story_detail.html', context)
 
-
+@login_required
 def add_stories(request):
     """ Add a story to customer stories"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you must have an account to do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = StoryForm(request.POST, request.FILES)
         if form.is_valid():
@@ -49,9 +54,12 @@ def add_stories(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_story(request, story_id):
     """ User to edit thier story """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you must have an account to do that.')
+        return redirect(reverse('home'))
 
     story = get_object_or_404(Stories, pk=story_id)
     if request.method == 'POST':
@@ -74,9 +82,12 @@ def edit_story(request, story_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_story(request, story_id):
     """ User to delate a story """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you must have an account to do that.')
+        return redirect(reverse('home'))
     story = get_object_or_404(Stories, pk=story_id)
     story.delete()
     messages.success(request, 'Your story is deleted!')
